truedata-optionchain-service/requirements.txt
fastapi
uvicorn
pandas
truedata-ws
python-dotenv
redis
requests
streamlit
plotly
streamlit-autorefresh
